requirements.txt
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
flasgger==0.9.7.1
Flask==3.1.0
idna==3.10
influxdb-client==1.48.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mistune==3.1.2
packaging==24.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
reactivex==4.0.4
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
setuptools==76.0.0
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
